refactor - refactoring ProductItemList, remove AsyncBoundary
