+++ pages/_document.tsx
@@ -62,11 +62,11 @@ export default class MyDocument extends Document {
           />
           <link
             href="https://fonts.googleapis.com/css2?family=Montserrat+Alternates:wght@400;500;600;700;800&display=swap"
-            rel="preload stylesheet"
+            rel="stylesheet"
             as="style"
           />
           <link
-            rel="preload"
+            rel="stylesheet"
             as="style"
             href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.4/dist/web/static/pretendard.css"
             crossOrigin=""
